+++ das/subscriber.go
@@ -32,6 +32,6 @@ func (s *subscriber) run(ctx context.Context, sub header.Subscription, emit list
 		}
 		log.Infow("new header received via subscription", "height", h.Height)
 
-		emit(ctx, uint64(h.Height))
+		emit(ctx, uint(h.Height))
 	}
 }
